Make orientation order explicit
